ocr-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
rapidocr-onnxruntime==1.3.24
Pillow==10.4.0
pydantic==2.9.2
